presentation of results

+++ include/timer.h
@@ -24,5 +24,5 @@ class Timer {
   uint64_t duration_since(uint64_t earlier);
   int get_hertz();
   double to_sec(uint64_t time);
-  uint64_t to_milli(uint64_t time);
+  double to_milli(uint64_t time);
 };
